Made the system asymmetrical and more robust

+++ Yeezus/src/yeezus/Main.java
@@ -2,7 +2,6 @@ package yeezus;
 
 import yeezus.driver.CPUSchedulingPolicy;
 import yeezus.driver.Driver;
-import yeezus.memory.MMU;
 import yeezus.memory.Memory;
 import yeezus.pcb.PCB;
 import yeezus.pcb.TaskManager;
@@ -19,60 +18,65 @@ public class Main {
 	private final static int CACHE_SIZE = 100;
 	private final static int REGISTER_SIZE = 16;
 
-	public static void main( String[] args )
-			throws Exception { // Since there shouldn't be an exception, we want everything to fail if there is one
-		// Initialize memory
-		Memory disk = new Memory( DISK_SIZE );
-		MMU mmu = new MMU( new Memory( RAM_SIZE ) );
+	public static void main( String[] args ) {
+		Memory disk;
+		Driver driver;
 
-		// Initialize Driver
-		Driver.loadFile( disk, new File( "src/yeezus/Program-File.txt" ) );
+		try {
+			// Initialize memory
+			disk = new Memory( DISK_SIZE );
 
-		Thread[] threads = new Thread[NUM_CPUS];
-
-		// Create Drivers
-		for ( int i = 0; i < NUM_CPUS; i++ ) {
-			Driver driver = new Driver( i, disk, mmu, REGISTER_SIZE, CACHE_SIZE, CPUSchedulingPolicy.FCFS );
-			Thread newThread = new Thread( driver );
-			threads[i] = newThread;
-			newThread.setUncaughtExceptionHandler( ( t, e ) -> {
-				e.printStackTrace();
-				System.err.println( driver.getCpu().generateDump() );
-			} );
+			// Initialize and create Driver
+			Driver.loadFile( disk, new File( "src/yeezus/Program-File.txt" ) );
+			driver = new Driver( NUM_CPUS, disk, REGISTER_SIZE, CACHE_SIZE, RAM_SIZE, CPUSchedulingPolicy.FCFS );
+		} catch ( Exception e ) {
+			System.err.println( "An exception occurred in system initialization." );
+			e.printStackTrace();
+			return;
 		}
 
 		// Log start time
 		long startTime = System.currentTimeMillis();
 
-		// Start threads
-		for ( int i = 0; i < NUM_CPUS; i++ ) {
-			threads[i].start();
+		try {
+			driver.run();
+		} catch ( Exception e ) {
+			System.err.println( "An exception occurred in system Execution." );
+			e.printStackTrace();
+			driver.dumpData();
+			return;
 		}
-
-		// (Busy) Wait for the threads
-		for ( Thread thread : threads ) {
-			thread.join( 5 );
-		}
-
 		// Log end time
 		long endTime = System.currentTimeMillis();
 
 		System.out.println( "The system completed in " + ( endTime - startTime ) + " milliseconds." );
 
 		// Print out the disk
-		File output = new File( "src/yeezus/Output_File.txt" );
-		output.createNewFile();
-		PrintStream out = new PrintStream( new FileOutputStream( output ) );
-		for ( int i = 0; i < disk.getCapacity(); i++ ) {
-			out.println( disk.read( i ) );
+		try {
+			File output = new File( "src/yeezus/Output_File.txt" );
+			if ( !output.exists() && !output.createNewFile() ) {
+				throw new Exception( "The output file could not be created." ); // idk how else to exit a try block
+			}
+			PrintStream out = new PrintStream( new FileOutputStream( output ) );
+			for ( int i = 0; i < disk.getCapacity(); i++ ) {
+				out.println( disk.read( i ) );
+			}
+		} catch ( Exception e ) {
+			System.err.println( "An exception occurred while writing to the output file." );
+			e.printStackTrace();
 		}
 
 		// Print out runtime information
-		for ( PCB pcb : TaskManager.INSTANCE.getPCBs() ) {
-			System.out.println(
-					"Process: " + pcb.getPID() + "\nWait Time: " + pcb.getElapsedWaitTime() + "\nRun Time: " + pcb
-							.getElapsedRunTime() + "\nExecution Count: " + pcb.getExecutionCount() + "\n" + "IO Count: "
-							+ pcb.getNumIO() + "\n" );
+		try {
+			for ( PCB pcb : TaskManager.INSTANCE.getPCBs() ) {
+				System.out.println(
+						"Process: " + pcb.getPID() + "\nWait Time: " + pcb.getElapsedWaitTime() + "\nRun Time: " + pcb
+								.getElapsedRunTime() + "\nExecution Count: " + pcb.getExecutionCount() + "\n"
+								+ "IO Count: " + pcb.getNumIO() + "\n" );
+			}
+		} catch ( Exception e ) {
+			System.err.println( "An exception occurred while printing the process data." );
+			e.printStackTrace();
 		}
 	}
 }
